Convert to a function to closer match python
